requirements.txt
feedparser
requests
newspaper3k
lxml_html_clean
python-dotenv
pandas
pyarrow
matplotlib
tqdm
polars
openpyxl